CareerCatalyst2/CareerCatalyst/backend/requirements.txt
fastapi
uvicorn
pydantic
spacy
scikit-learn
requests
python-docx
pdfplumber 
google-generativeai
flask
flask-cors
python-dotenv 